src/evaluator.py
from typing import Dict, Any
import yaml
import os
from dotenv import load_dotenv

load_dotenv()

class Evaluator:
    def __init__(self, config_path: str = "config.yaml"):
        try:
            with open(config_path, encoding='utf-8') as f:
                config_content = f.read()
                config_content = os.path.expandvars(config_content)
                self.config = yaml.safe_load(config_content)
        except FileNotFoundError:
            raise FileNotFoundError(f"設定ファイル {config_path} が見つかりません")
        except UnicodeDecodeError:
            raise UnicodeDecodeError("設定ファイルの文字コードが不正です。UTF-8で保存してください")
        except yaml.YAMLError as e:
            raise ValueError(f"設定ファイルのYAML形式が不正です: {str(e)}")
        
    def evaluate(self, analysis_result: Dict) -> tuple[float, bool]:
        """分析結果を評価"""
        # TODO: 実際の評価ロジックを実装
        score = 0.85  # 仮のスコア
        needs_improvement = score < 0.9  # スコアが0.9未満なら改善が必要
        return score, needs_improvement
