pyproject.toml
[build-system]
requires = ["setuptools>=65", "uv>=0.7.13"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-investment-assistant"
version = "0.1.0"
description = "AI-powered stock investment assistant tool"
authors = [
    { name = "Fumiyoshi To", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pandas==2.3.0",
    "numpy==2.2.5",
    "requests==2.32.3",
    "python-dotenv==1.1.0",
    "PyYAML==6.0.2",
    "mcp[cli]>=1.9.4",
    "mcp-agent>=0.0.1",
]
